Save plaers

+++ Final-project/src/main/java/TicTacToe/repository/PlayerRep.java
@@ -3,5 +3,5 @@ package TicTacToe.repository;
 import TicTacToe.model.Player;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PlayerRep extends CrudRepository<Player, Long> {
+public interface PlayerRep extends CrudRepository<Player, Integer> {
 }
